pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repovideo"
version = "0.1.0"
description = "Take any repo, run it, and generate a polished demo video with AI-generated intro"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "gitpython>=3.1",
    "docker>=7.0",
    "ffmpeg-python>=0.2",
    "playwright>=1.40",
    "Pillow>=10.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "litellm>=1.30",
    "diffusers>=0.30",
    "transformers>=4.40",
    "accelerate>=0.30",
    "peft>=0.11",
    "torch>=2.2",
    "safetensors>=0.4",
    "sentencepiece>=0.2",
    "datasets>=2.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
repovideo = "src.cli:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100
